cleanup: Got rid of ADDON_SCRIPT_LIBRARY, compile warnings, and cosmetics.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@30493 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/GUIDialogAddonInfo.cpp
@@ -241,7 +241,7 @@ void CGUIDialogAddonInfo::OnJobComplete(unsigned int jobID, bool success,
     if (file.Open("special://temp/"+
       CUtil::GetFileName(((CFileOperationJob*)job)->GetItems()[0]->m_strPath)))
     {
-      char* temp = new char[file.GetLength()+1];
+      char* temp = new char[(size_t)file.GetLength()+1];
       file.Read(temp,file.GetLength());
       temp[file.GetLength()] = '\0';
       m_item->SetProperty("Addon.Changelog",temp);
